requirements.txt
beautifulsoup4
google-api-python-client
flask
redis
gunicorn
pytesseract
gingerit
pillow
nltk
requests
PyMuPDF
gensim
Werkzeug
fuzzywuzzy
python-docx
python-levenshtein
rq
textract
